Move quadKey function to the one place it is used
