+++ src/ol/tilecoord.js
@@ -71,31 +71,6 @@ export function hash(tileCoord) {
 
 
 /**
- * @param {TileCoord} tileCoord Tile coord.
- * @return {string} Quad key.
- */
-export function quadKey(tileCoord) {
-  const z = tileCoord[0];
-  const digits = new Array(z);
-  let mask = 1 << (z - 1);
-  let i, charCode;
-  for (i = 0; i < z; ++i) {
-    // 48 is charCode for 0 - '0'.charCodeAt(0)
-    charCode = 48;
-    if (tileCoord[1] & mask) {
-      charCode += 1;
-    }
-    if (tileCoord[2] & mask) {
-      charCode += 2;
-    }
-    digits[i] = String.fromCharCode(charCode);
-    mask >>= 1;
-  }
-  return digits.join('');
-}
-
-
-/**
  * @param {TileCoord} tileCoord Tile coordinate.
  * @param {!import("./tilegrid/TileGrid.js").default} tileGrid Tile grid.
  * @return {boolean} Tile coordinate is within extent and zoom level range.
